fix renaming conflicts
