+++ src/main/java/seedu/smartnus/logic/parser/SmartNusParser.java
@@ -13,6 +13,7 @@ import seedu.smartnus.logic.commands.Command;
 import seedu.smartnus.logic.commands.DeleteCommand;
 import seedu.smartnus.logic.commands.EditCommand;
 import seedu.smartnus.logic.commands.ExitCommand;
+import seedu.smartnus.logic.commands.FindByTagCommand;
 import seedu.smartnus.logic.commands.FindCommand;
 import seedu.smartnus.logic.commands.HelpCommand;
 import seedu.smartnus.logic.commands.ListCommand;
