+++ src/main/java/seedu/smartnus/model/question/ShowAllQuestionsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.question;
+package seedu.smartnus.model.question;
 
 import java.util.function.Predicate;
